Update sunset_policies.md
Hello! A customer was confused about the current documentation because I had told him we automatically suppress users who have reported email as spam, and he pointed out that we specifically say to do that here. I updated some of the verbiage to (hopefully) be more clear! We may want tech support to review, but my understanding is that it's easier in Braze for customers to target based on prior opens/clicks rather than bulk-suppress users who have *not* engaged so I added a bit about that, too.  Thank you!!